requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
PyJWT==2.3.0
python-dotenv==0.19.0
openai==0.27.0
requests==2.28.1
google-genai
pillow
pinecone
langchain-pinecone
langchain-openai
langchain-text-splitters
langchain
langchain-community
pypdf
gitingest
openai-agents
langchain-permit